FullPrunedBlockChain,FullPrunedBlockStore,MemoryFullPrunedBlockStore: move to integration-test
