+++ integration-test/build.gradle
@@ -9,6 +9,7 @@ dependencies {
     implementation project(':bitcoinj-core')
     implementation project(':bitcoinj-examples')
     api 'org.jspecify:jspecify:1.0.1'
+    implementation 'org.slf4j:slf4j-api:2.0.18'
 
     testImplementation 'org.slf4j:slf4j-jdk14:2.0.18'
     testImplementation platform('org.junit:junit-bom:6.1.3')
