+++ integration-test/src/main/java/org/bitcoinj/node/store/FullPrunedBlockStore.java
@@ -14,13 +14,15 @@
  * limitations under the License.
  */
 
-package org.bitcoinj.store;
+package org.bitcoinj.node.store;
 
 import org.bitcoinj.base.Sha256Hash;
 import org.bitcoinj.core.StoredBlock;
 import org.bitcoinj.core.StoredUndoableBlock;
 import org.bitcoinj.core.UTXO;
 import org.bitcoinj.core.UTXOProvider;
+import org.bitcoinj.store.BlockStore;
+import org.bitcoinj.store.BlockStoreException;
 import org.jspecify.annotations.Nullable;
