+++ tools/src/main/java/org/bitcoinj/tools/BlockImporter.java
@@ -20,15 +20,15 @@ import org.bitcoinj.base.BitcoinNetwork;
 import org.bitcoinj.core.AbstractBlockChain;
 import org.bitcoinj.core.Block;
 import org.bitcoinj.core.BlockChain;
-import org.bitcoinj.core.FullPrunedBlockChain;
+import org.bitcoinj.node.chain.FullPrunedBlockChain;
 import org.bitcoinj.core.NetworkParameters;
 import org.bitcoinj.core.PrunedException;
 import org.bitcoinj.core.VerificationException;
+import org.bitcoinj.node.store.MemoryFullPrunedBlockStore;
 import org.bitcoinj.store.BlockStore;
 import org.bitcoinj.store.BlockStoreException;
-import org.bitcoinj.store.FullPrunedBlockStore;
+import org.bitcoinj.node.store.FullPrunedBlockStore;
 import org.bitcoinj.store.MemoryBlockStore;
-import org.bitcoinj.store.MemoryFullPrunedBlockStore;
 import org.bitcoinj.store.SPVBlockStore;
 import org.bitcoinj.utils.BlockFileLoader;
